refactor : 사용자 input 유효성 검사하는 코드 분리
